[RELEASE] Android v1.0.0 병합

+++ android/app/src/main/java/com/daedan/festabook/presentation/setting/SettingViewModel.kt
@@ -10,63 +10,92 @@ import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
 import com.daedan.festabook.FestaBookApp
 import com.daedan.festabook.domain.repository.FestivalNotificationRepository
-import com.daedan.festabook.presentation.common.Event
+import com.daedan.festabook.presentation.common.SingleLiveData
 import kotlinx.coroutines.launch
 import timber.log.Timber
 
 class SettingViewModel(
     private val festivalNotificationRepository: FestivalNotificationRepository,
 ) : ViewModel() {
-    private val _allowClickEvent: MutableLiveData<Event<Unit>> = MutableLiveData()
-    val allowClickEvent: LiveData<Event<Unit>> get() = _allowClickEvent
+    private val _permissionCheckEvent: SingleLiveData<Unit> = SingleLiveData()
+    val permissionCheckEvent: LiveData<Unit> get() = _permissionCheckEvent
 
-    var isAllowed = festivalNotificationRepository.getFestivalNotificationIsAllow()
-        private set
+    private val _isAllowed =
+        MutableLiveData(
+            festivalNotificationRepository.getFestivalNotificationIsAllow(),
+        )
+    val isAllowed: LiveData<Boolean> get() = _isAllowed
 
-    private val _error: MutableLiveData<Event<Throwable>> = MutableLiveData()
-    val error: LiveData<Event<Throwable>> get() = _error
+    private val _error: SingleLiveData<Throwable> = SingleLiveData()
+    val error: LiveData<Throwable> get() = _error
+
+    private val _isLoading: MutableLiveData<Boolean> = MutableLiveData(false)
+    val isLoading: LiveData<Boolean> get() = _isLoading
+
+    private val _success: SingleLiveData<Unit> = SingleLiveData()
+    val success: LiveData<Unit> get() = _success
 
     fun notificationAllowClick() {
-        updateNotificationIsAllowed(!isAllowed)
-        _allowClickEvent.value = Event(Unit)
-        Timber.d("$isAllowed")
-        if (isAllowed) saveNotificationId() else deleteNotificationId()
+        if (_isAllowed.value == false) {
+            _permissionCheckEvent.setValue(Unit)
+        } else {
+            deleteNotificationId()
+        }
     }
 
-    fun saveNotificationIsAllowed(isAllowed: Boolean) {
+    private fun saveNotificationIsAllowed(isAllowed: Boolean) {
         festivalNotificationRepository.setFestivalNotificationIsAllow(isAllowed)
     }
 
-    fun updateNotificationIsAllowed(allowed: Boolean) {
-        isAllowed = allowed
+    private fun updateNotificationIsAllowed(allowed: Boolean) {
+        _isAllowed.value = allowed
     }
 
-    private fun saveNotificationId() {
+    fun saveNotificationId() {
+        if (_isLoading.value == true) return
+        _isLoading.value = true
+
+        // Optimistic UI 적용, 요청 실패 시 원복
+        saveNotificationIsAllowed(true)
+        updateNotificationIsAllowed(true)
+        _success.setValue(Unit)
+
         viewModelScope.launch {
             val result =
                 festivalNotificationRepository.saveFestivalNotification()
 
             result
-                .onSuccess {
-                    saveNotificationIsAllowed(isAllowed)
-                }.onFailure {
-                    _error.value = Event(it)
-                    Timber.e(it, "${::SettingViewModel.javaClass.simpleName} NotificationId 저장 실패")
+                .onFailure {
+                    _error.setValue(it)
+                    saveNotificationIsAllowed(false)
+                    updateNotificationIsAllowed(false)
+                    Timber.e(it, "${this::class.java.simpleName} NotificationId 저장 실패")
+                }.also {
+                    _isLoading.value = false
                 }
         }
     }
 
     private fun deleteNotificationId() {
+        if (_isLoading.value == true) return
+        _isLoading.value = true
+
+        // Optimistic UI 적용, 요청 실패 시 원복
+        saveNotificationIsAllowed(false)
+        updateNotificationIsAllowed(false)
+
         viewModelScope.launch {
             val result =
                 festivalNotificationRepository.deleteFestivalNotification()
 
             result
-                .onSuccess {
-                    saveNotificationIsAllowed(isAllowed)
-                }.onFailure {
-                    _error.value = Event(it)
-                    Timber.e(it, "${::SettingViewModel.name} NotificationId 삭제 실패")
+                .onFailure {
+                    _error.setValue(it)
+                    saveNotificationIsAllowed(true)
+                    updateNotificationIsAllowed(true)
+                    Timber.e(it, "${this::class.java.simpleName} NotificationId 삭제 실패")
+                }.also {
+                    _isLoading.value = false
                 }
         }
     }
